Calculator
A simple calculator application that asks a user to enter two numbers and the operation (eg. +, -, x, etc.) that they'd like to perform on the numbers. Displays the answer to the equation. Every equation entered by the user should be written to a text file. 

Program gives the user the option to either enter two numbers and an operator, or read all the equations from a new text file (the user should add the name of the txt file as an input) and print out all the equations together with the results.